Be elaborate when specifying the debs on the download page.
Junio hasn't done the debian packages for git in a while, and as
such the location specified on the site is outdated. Incidentally
git-core _is_ packaged for debian and versions of it are in unstable
and testing. There is a version for stable maintained at backports.org.

This commit modifies the downloads page to reflect these changes.

Signed-Off-By: Martin Atukunda <matlads@dsmagic.com>