Fold all the stuff to a single page
When designing the multi-page layout I expected the homepage to bear
more information but that didn't turn out to be the case and now I'll
rather leave the job to the wiki (and patch submitters ;). So per Linus
suggestion fold everything to a single page and make the menu bar actually
useful for quickly going to other Git-related sites.