download.html: Link to both the gitweb and GitHub web interface
The repo.or.cz gitweb interface is linked because of the graphiclog
and forking capability, on par with GitHub features; also, simply many
existing forks of git.git are there already that the readers may find
interesting.

There is already a link to gitweb interface in the download box but
I think the extra redundancy here does no harm, since the reader might
not think about looking away to a box elsewhere on the page when reading
about how to access the Git's repository.